Respect target specificity for configuration instead of just picking js/ts over typescript.
